requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.2

# Data Processing
pandas==2.1.4
numpy==1.26.2
yfinance==0.2.35
scikit-learn==1.3.2
tensorflow>=2.16.0
matplotlib==3.8.2
seaborn==0.13.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Zerodha API
kiteconnect==4.1.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
apscheduler==3.10.4
requests==2.31.0 